chore: improve styles in `header` component
